fix: #347
